Rename : 코드의 이름 변경
이름의 형식이 맞지않거나, url의 형식을 맞춰주기 위해 이름을 변경
