+++ src/main/resources/templates/navigation_bar.html
@@ -67,7 +67,7 @@
                 <ul class="nav navbar-nav navbar-right">
                     <li><a th:href="@{/}">Posts</a></li>
                     <li th:if="${session.login == null}"><a th:href="@{/user/login}" role="button">로그인</a></li>
-                    <li th:if="${session.login == null}"><a th:href="@{/user/createForm}" role="button">회원가입</a></li>
+                    <li th:if="${session.login == null}"><a th:href="@{/user/create}" role="button">회원가입</a></li>
                     <li th:if="${session.login == true}"><a th:href="@{/user/logout}" role="button">로그아웃</a></li>
                     <li th:if="${session.login == true}"><a th:href="@{/user/update}"role="button">개인정보수정</a></li>
                 </ul>
